Add ability to set `--customPlatform` flag
Kaniko supports a `--customPlatform` flag to allow creation of images
with a different platform than the default, which is a platform derived
from the build host.

The `--customPlatform` feature doesn't provide any cross-compilation
support by itself, and only platforms which the underlying build host
is able to provide will be supported.

This commit provides a `custom_platform:` plugin parameter which,
when supplied, will be provide a value for the `--customPlatform`
argument for `/kaniko/executor`. There is no attempt to verify that
the platform requested is supported by the build host.

Example:

```
steps:
  - name: publish_hello-world_amd64
    image: vela-plugins/kaniko:vCustomPlatform
    pull: always
    parameters:
      repo: vela/hello-world
      custom_platform: linux/amd64
      tags:
        - latest_linux_amd64
  - name: publish_hello-world_arm64
    image: vela-plugins/kaniko:vCustomPlatform
    pull: always
    parameters:
      repo: vela/hello-world
      custom_platform: linux/arm64/v8
      tags:
        - latest_linux_arm64_v8
```